chore(UI): rename yt api init [YTFRONT-4982]
